Removed metricService
