+++ public/controllers/overview.js
@@ -1,8 +1,8 @@
 const app = require('ui/modules').get('app/wazuh', []);
 import $ from 'jquery';
 
-app.controller('overviewController', function ($scope, $location, $rootScope, appState, genericReq, errorHandler, metricService) {
-
+app.controller('overviewController', function ($scope, $location, $rootScope, appState, genericReq, errorHandler) {
+    $rootScope.rawVisualizations = null;
     // Timestamp for visualizations at controller's startup
     if(!$rootScope.visTimestamp) {
         $rootScope.visTimestamp = new Date().getTime();
@@ -110,31 +110,53 @@ app.controller('overviewController', function ($scope, $location, $rootScope, ap
         virustotal: { group: 'virustotal' }
     };
 
-    const checkMetrics = (tab, subtab) => {
-        metricService.destroyWatchers();
+    const generateMetric = id => {
+        let html = $(id).html();
+        if (typeof html !== 'undefined' && html.includes('<span')) {
+            if(typeof html.split('<span>')[1] !== 'undefined'){
+                return html.split('<span>')[1].split('</span')[0];
+            } else if(html.includes('table') && html.includes('cell-hover')){
+                let nonB = html.split('ng-non-bindable')[1];
+                if(nonB && 
+                    nonB.split('>')[1] && 
+                    nonB.split('>')[1].split('</')[0]
+                ) {
+                    return nonB.split('>')[1].split('</')[0];
+                }
+            }
+        }
+        return '';
+    }
+    
+    const createMetrics = metricsObject => {
+        for(let key in metricsObject) {
+            $scope[key] = () => generateMetric(metricsObject[key]);
+        }
+    }
 
+    const checkMetrics = (tab, subtab) => {
         if(subtab === 'panels'){
             switch (tab) {
                 case 'general':
-                    metricService.createWatchers(metricsGeneral);
+                    createMetrics(metricsGeneral);
                     break;
                 case 'fim':
-                    metricService.createWatchers(metricsFim);
+                    createMetrics(metricsFim);
                     break;
                 case 'audit':
-                    metricService.createWatchers(metricsAudit);
+                    createMetrics(metricsAudit);
                     break;
                 case 'vuls':
-                    metricService.createWatchers(metricsVulnerability);
+                    createMetrics(metricsVulnerability);
                     break;
                 case 'oscap':
-                    metricService.createWatchers(metricsScap);
+                    createMetrics(metricsScap);
                     break;
                 case 'virustotal':
-                    metricService.createWatchers(metricsVirustotal);
+                    createMetrics(metricsVirustotal);
                     break;
                 case 'aws':
-                    metricService.createWatchers(metricsAws);
+                    createMetrics(metricsAws);
                     break;
             }
         }
@@ -232,8 +254,6 @@ app.controller('overviewController', function ($scope, $location, $rootScope, ap
                 h._scope.$destroy();
             }
         }
-
-        if(metricService.hasItems()) metricService.destroyWatchers();
 
         $rootScope.ownHandlers = [];
     });
