+++ public/templates/agents/agents-audit.html
@@ -5,10 +5,10 @@
         <div layout="row">
             <md-card flex class="wz-metric-color wz-md-card">
                 <md-card-content layout="row" class="wz-padding-metric">
-                    <div class="wz-text-truncatable" flex>New files: <span class="wz-text-bold">{{auditNewFiles}}</span></div>
-                    <div class="wz-text-truncatable" flex>Read files: <span class="wz-text-bold">{{auditReadFiles}}</span></div>
-                    <div class="wz-text-truncatable" flex>Modified files: <span class="wz-text-bold">{{auditModifiedFiles}}</span></div>
-                    <div class="wz-text-truncatable" flex>Removed files: <span class="wz-text-bold">{{auditRemovedFiles}}</span></div>
+                    <div class="wz-text-truncatable" flex>New files: <span class="wz-text-bold" ng-bind="auditNewFiles()"></span></div>
+                    <div class="wz-text-truncatable" flex>Read files: <span class="wz-text-bold" ng-bind="auditReadFiles()"></span></div>
+                    <div class="wz-text-truncatable" flex>Modified files: <span class="wz-text-bold" ng-bind="auditModifiedFiles()"></span></div>
+                    <div class="wz-text-truncatable" flex>Removed files: <span class="wz-text-bold" ng-bind="auditRemovedFiles()"></span></div>
                 </md-card-content>
             </md-card>
         </div>
