+++ public/templates/agents/agents-oscap.html
@@ -5,9 +5,9 @@
         <div layout="row">
             <md-card flex class="wz-metric-color wz-md-card">
                 <md-card-content layout="row" class="wz-padding-metric">
-                    <div flex>Last score: <span class="wz-text-bold">{{scapLastScore}}</span></div>
-                    <div flex>Highest score: <span class="wz-text-bold">{{scapHighestScore}}</span></div>
-                    <div flex>Lowest score: <span class="wz-text-bold">{{scapLowestScore}}</span></div>
+                    <div flex>Last score: <span class="wz-text-bold" ng-bind="scapLastScore()"></span></div>
+                    <div flex>Highest score: <span class="wz-text-bold" ng-bind="scapHighestScore()"></span></div>
+                    <div flex>Lowest score: <span class="wz-text-bold" ng-bind="scapLowestScore()"></span></div>
                 </md-card-content>
             </md-card>
         </div>
