+++ public/templates/agents/agents-virustotal.html
@@ -6,9 +6,9 @@
         <div layout="row">
             <md-card flex class="wz-metric-color wz-md-card">
                 <md-card-content layout="row" class="wz-padding-metric">
-                    <div flex>Total malicious: <span class="wz-text-bold">{{virusMalicious}}</span></div>
-                    <div flex>Total positives: <span class="wz-text-bold">{{virusPositives}}</span></div>
-                    <div flex>Total: <span class="wz-text-bold">{{virusTotal}}</span></div>                    
+                    <div flex>Total malicious: <span class="wz-text-bold" ng-bind="virusMalicious()"></span></div>
+                    <div flex>Total positives: <span class="wz-text-bold" ng-bind="virusPositives()"></span></div>
+                    <div flex>Total: <span class="wz-text-bold" ng-bind="virusTotal()"></span></div>                    
                 </md-card-content>
             </md-card>
         </div>
