+++ public/templates/agents/agents-vuls.html
@@ -6,10 +6,10 @@
         <div layout="row">
             <md-card flex class="wz-metric-color wz-md-card">
                 <md-card-content layout="row" class="wz-padding-metric">
-                    <div flex>Critical severity alerts: <span class="wz-text-bold">{{vulnCritical}}</span></div>
-                    <div flex>High severity alerts: <span class="wz-text-bold">{{vulnHigh}}</span></div>
-                    <div flex>Medium severity alerts: <span class="wz-text-bold">{{vulnMedium}}</span></div>
-                    <div flex>Low severity alerts: <span class="wz-text-bold">{{vulnLow}}</span></div>
+                    <div flex>Critical severity alerts: <span class="wz-text-bold" ng-bind="vulnCritical()"></span></div>
+                    <div flex>High severity alerts: <span class="wz-text-bold" ng-bind="vulnHigh()"></span></div>
+                    <div flex>Medium severity alerts: <span class="wz-text-bold" ng-bind="vulnMedium()"></span></div>
+                    <div flex>Low severity alerts: <span class="wz-text-bold" ng-bind="vulnLow()"></span></div>
                 </md-card-content>
             </md-card>
         </div>
