+++ public/templates/overview/overview-aws.html
@@ -3,10 +3,10 @@
     <div layout="row">
         <md-card flex class="wz-metric-color wz-md-card">
             <md-card-content layout="row" class="wz-padding-metric">
-                <div class="wz-text-truncatable" flex>Successful logins: <span class="wz-text-bold">{{awsLogins}}</span></div>
-                <div class="wz-text-truncatable" flex>Most active user: <span class="wz-text-bold">{{awsMostActiveUser}}</span></div>
-                <div class="wz-text-truncatable" flex>Authorized security groups: <span class="wz-text-bold">{{awsAuthorized}}</span></div>    
-                <div class="wz-text-truncatable" flex>Revoked security groups: <span class="wz-text-bold">{{awsRevoked}}</span></div>
+                <div class="wz-text-truncatable" flex>Successful logins: <span class="wz-text-bold" ng-bind="awsLogins()"></span></div>
+                <div class="wz-text-truncatable" flex>Most active user: <span class="wz-text-bold" ng-bind="awsMostActiveUser()"></span></div>
+                <div class="wz-text-truncatable" flex>Authorized security groups: <span class="wz-text-bold" ng-bind="awsAuthorized()"></span></div>    
+                <div class="wz-text-truncatable" flex>Revoked security groups: <span class="wz-text-bold" ng-bind="awsRevoked()"></span></div>
             </md-card-content>
         </md-card>
     </div>
