+++ public/templates/overview/overview-fim.html
@@ -3,9 +3,9 @@
     <div layout="row">
         <md-card flex class="wz-metric-color wz-md-card">
             <md-card-content layout="row" class="wz-padding-metric">
-                <div class="wz-text-truncatable" flex>Files added: <span class="wz-text-bold">{{fimAdded}}</span></div>
-                <div class="wz-text-truncatable" flex>Files modified: <span class="wz-text-bold">{{fimModified}}</span></div>
-                <div class="wz-text-truncatable" flex>Files deleted: <span class="wz-text-bold">{{fimDeleted}}</span></div>
+                <div class="wz-text-truncatable" flex>Files added: <span class="wz-text-bold" ng-bind="fimAdded()"></span></div>
+                <div class="wz-text-truncatable" flex>Files modified: <span class="wz-text-bold" ng-bind="fimModified()"></span></div>
+                <div class="wz-text-truncatable" flex>Files deleted: <span class="wz-text-bold" ng-bind="fimDeleted()"></span></div>
             </md-card-content>
         </md-card>
     </div>
