+++ public/templates/overview/overview-general.html
@@ -3,10 +3,10 @@
     <div layout="row">
         <md-card flex class="wz-metric-color wz-md-card">
             <md-card-content layout="row" class="wz-padding-metric">
-                <div class="wz-text-truncatable" flex>Alerts: <span class="wz-text-bold">{{totalAlerts}}</span></div>
-                <div class="wz-text-truncatable" flex>Level 12 or above alerts: <span class="wz-text-bold">{{level12}}</span></div>
-                <div class="wz-text-truncatable" flex>Authentication failure: <span class="wz-text-bold">{{authFailure}}</span></div>
-                <div class="wz-text-truncatable" flex>Authentication success: <span class="wz-text-bold">{{authSuccess}}</span></div>
+                <div class="wz-text-truncatable" flex>Alerts: <span class="wz-text-bold" ng-bind="totalAlerts()"></span></div>
+                <div class="wz-text-truncatable" flex>Level 12 or above alerts: <span class="wz-text-bold" ng-bind="level12()"></span></div>
+                <div class="wz-text-truncatable" flex>Authentication failure: <span class="wz-text-bold" ng-bind="authFailure()"></span></div>
+                <div class="wz-text-truncatable" flex>Authentication success: <span class="wz-text-bold" ng-bind="authSuccess()"></span></div>
             </md-card-content>
         </md-card>
     </div>
