+++ public/templates/overview/overview-oscap.html
@@ -3,9 +3,9 @@
     <div layout="row">
         <md-card flex class="wz-metric-color wz-md-card">
             <md-card-content layout="row" class="wz-padding-metric">
-                <div class="wz-text-truncatable" flex>Last score: <span class="wz-text-bold">{{scapLastScore}}</span></div>
-                <div class="wz-text-truncatable" flex>Highest score: <span class="wz-text-bold">{{scapHighestScore}}</span></div>
-                <div class="wz-text-truncatable" flex>Lowest score: <span class="wz-text-bold">{{scapLowestScore}}</span></div>
+                <div class="wz-text-truncatable" flex>Last score: <span class="wz-text-bold" ng-bind="scapLastScore()"></span></div>
+                <div class="wz-text-truncatable" flex>Highest score: <span class="wz-text-bold" ng-bind="scapHighestScore()"></span></div>
+                <div class="wz-text-truncatable" flex>Lowest score: <span class="wz-text-bold" ng-bind="scapLowestScore()"></span></div>
             </md-card-content>
         </md-card>
     </div>
